requirements.txt
flask
ultralytics
opencv-python
torch
torchvision
pillow
numpy
matplotlib
requests
